modules/api_kafka/requirements.txt
grpcio-tools
protobuf
kafka-python
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
Flask-SQLAlchemy >= 2.5.1
shapely==1.7.0
psycopg2-binary